Do not build internal metrics without CGO (#6501)
* libbeat: build internal metrics on windows and linux without cgo

* libbeat: add ephemeral id even if no internal memory metrics is reported

* fix incorrect build tags && create common file for all platforms
