+++ libbeat/cmd/instance/metrics_other.go
@@ -1,8 +1,12 @@
-// +build !darwin,!linux,!windows darwin,!cgo linux,!cgo windows,!cgo
+// +build !darwin !cgo
+// +build !freebsd !cgo
+// +build !linux,!windows
 
 package instance
 
-import "github.com/elastic/beats/libbeat/logp"
+import (
+	"github.com/elastic/beats/libbeat/logp"
+)
 
 func setupMetrics(name string) error {
 	logp.Warn("Metrics not implemented for this OS.")
